+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Renamed `ExecutorTask` -> `AsyncTask`, return value to `FutureOr`.
 - Renamed `Rate.limit` -> `Rate.maximum`.
 
+**New features**
+
+- Added `Executor.onChange`. Clients can use this to monitor the current `scheduledCount` and queue more tasks to ensure `Executor` is running on full capacity.
+
 ## 0.1.2
 
 - Expose internal stats: `runningCount`, `waitingCount`, `scheduledCount`.